All parts finished.Needs controls to accomplish specific I/O's

+++ ft_strlcat.c
@@ -6,8 +6,8 @@ size_t ft_strlcat(char  *dst, const char    *src, size_t    n)
         return (0);
     
     size_t i;
-    int dst_len;
-    int src_len;
+    size_t dst_len;
+    size_t src_len;
 
     i = 0;
     dst_len = ft_strlen(dst);
